requirements.txt
flask
flask-cors
google-generativeai
gunicorn
waitress